16bits node-id

+++ examples/rpl_udp/viz.h
@@ -1,12 +1,12 @@
 /* CA: prototypes for viz.c */
 
-void viz_udp_pkt(uint8_t src);
+void viz_udp_pkt(uint16_t src);
 
-void viz_udp_snd(uint8_t dst);
+void viz_udp_snd(uint16_t dst);
 
-void viz_parent_select(uint8_t parent);
+void viz_parent_select(uint16_t parent);
 
-void viz_parent_deselect(uint8_t parent);
+void viz_parent_deselect(uint16_t parent);
 
-void viz_dio(uint8_t src);
+void viz_dio(uint16_t src);
